+++ export/cariva-ds-dev-export/ds/components/CrvButtonDecorative/index.ts
@@ -1,0 +1,5 @@
+export { CrvButtonDecorative } from './CrvButtonDecorative';
+export type {
+  CrvButtonDecorativeProps,
+  CrvButtonDecorativeSize,
+} from './CrvButtonDecorative.types';